improve import and move batchData to own module
