+++ tidy3d/web/api/asynchronous.py
@@ -7,7 +7,8 @@ from typing import Literal, Optional, Union
 from tidy3d.log import log
 from tidy3d.web.core.types import PayType
 
-from .container import DEFAULT_DATA_DIR, Batch, BatchData
+from .batch_data import BatchData
+from .container import DEFAULT_DATA_DIR, Batch
 from .tidy3d_stub import SimulationType
