+++ tidy3d/web/api/autograd/autograd.py
@@ -25,7 +25,8 @@ from tidy3d.components.data.data_array import DataArray
 from tidy3d.exceptions import AdjointError
 from tidy3d.web.api.asynchronous import DEFAULT_DATA_DIR
 from tidy3d.web.api.asynchronous import run_async as run_async_webapi
-from tidy3d.web.api.container import DEFAULT_DATA_PATH, Batch, BatchData, Job
+from tidy3d.web.api.batch_data import BatchData
+from tidy3d.web.api.container import DEFAULT_DATA_PATH, Batch, Job
 from tidy3d.web.api.tidy3d_stub import SimulationDataType, SimulationType
 from tidy3d.web.api.webapi import run as run_webapi
 from tidy3d.web.core.s3utils import download_file, upload_file
