+++ tidy3d/web/api/container.py
@@ -6,9 +6,8 @@ import concurrent
 import os
 import time
 from abc import ABC
-from collections.abc import Mapping
 from concurrent.futures import ThreadPoolExecutor
-from typing import Literal, Optional
+from typing import TYPE_CHECKING, Literal, Optional
 
 import pydantic.v1 as pd
 from rich.progress import BarColumn, Progress, TaskProgressColumn, TextColumn, TimeElapsedColumn
@@ -25,6 +24,9 @@ from tidy3d.web.core.task_info import RunInfo, TaskInfo
 from tidy3d.web.core.types import PayType
 
 from .tidy3d_stub import SimulationDataType, SimulationType
+
+if TYPE_CHECKING:
+    from tidy3d.web.api.batch_data import BatchData
 
 # Max # of workers for parallel upload / download: above 10, performance is same but with warnings
 DEFAULT_NUM_WORKERS = 10
@@ -433,89 +435,6 @@ class Job(WebContainer):
             os.makedirs(parent_dir, exist_ok=True)
 
 
-class BatchData(Tidy3dBaseModel, Mapping):
-    """
-    Holds a collection of :class:`.SimulationData` returned by :class:`Batch`.
-
-    Notes
-    -----
-
-        When the batch is completed, the output is not a :class:`.SimulationData` but rather a :class:`BatchData`. The
-        data within this :class:`BatchData` object can either be indexed directly ``batch_results[task_name]`` or can be looped
-        through ``batch_results.items()`` to get the :class:`.SimulationData` for each task.
-
-    See Also
-    --------
-
-    :class:`Batch`:
-         Interface for submitting several :class:`.Simulation` objects to sever.
-
-    :class:`.SimulationData`:
-         Stores data from a collection of :class:`.Monitor` objects in a :class:`.Simulation`.
-
-    **Notebooks**
-        * `Running simulations through the cloud <../../notebooks/WebAPI.html>`_
-        * `Performing parallel / batch processing of simulations <../../notebooks/ParameterScan.html>`_
-    """
-
-    task_paths: dict[TaskName, str] = pd.Field(
-        ...,
-        title="Data Paths",
-        description="Mapping of task_name to path to corresponding data for each task in batch.",
-    )
-
-    task_ids: dict[TaskName, str] = pd.Field(
-        ..., title="Task IDs", description="Mapping of task_name to task_id for each task in batch."
-    )
-
-    verbose: bool = pd.Field(
-        True, title="Verbose", description="Whether to print info messages and progressbars."
-    )
-
-    def load_sim_data(self, task_name: str) -> SimulationDataType:
-        """Load a simulation data object from file by task name."""
-        task_data_path = self.task_paths[task_name]
-        task_id = self.task_ids[task_name]
-        web.get_info(task_id)
-
-        return web.load(task_id=task_id, path=task_data_path, verbose=False)
-
-    def __getitem__(self, task_name: TaskName) -> SimulationDataType:
-        """Get the simulation data object for a given ``task_name``."""
-        return self.load_sim_data(task_name)
-
-    def __iter__(self):
-        """Iterate over the task names."""
-        return iter(self.task_paths)
-
-    def __len__(self):
-        """Return the number of tasks in the batch."""
-        return len(self.task_paths)
-
-    @classmethod
-    def load(cls, path_dir: str = DEFAULT_DATA_DIR, replace_existing: bool = False) -> BatchData:
-        """Load :class:`Batch` from file, download results, and load them.
-
-        Parameters
-        ----------
-        path_dir : str = './'
-            Base directory where data will be downloaded, by default current working directory.
-            A `batch.hdf5` file must be present in the directory.
-        replace_existing : bool = False
-            Downloads the data even if path exists (overwriting the existing).
-
-        Returns
-        ------
-        :class:`BatchData`
-            Contains Union[:class:`.SimulationData`, :class:`.HeatSimulationData`, :class:`.EMESimulationData`]
-            for each Union[:class:`.Simulation`, :class:`.HeatSimulation`, :class:`.EMESimulation`] in :class:`Batch`.
-        """
-
-        batch_file = Batch._batch_path(path_dir=path_dir)
-        batch = Batch.from_file(batch_file)
-        return batch.load(path_dir=path_dir, replace_existing=replace_existing)
-
-
 class Batch(WebContainer):
     """
     Interface for submitting several :class:`Simulation` objects to sever.
@@ -664,6 +583,9 @@ class Batch(WebContainer):
         ``max_num_adjoint_per_fwd`` fields.
         """
         self._check_path_dir(path_dir)
+
+        # Import BatchData here to avoid circular imports
+        from tidy3d.web.api.batch_data import BatchData
 
         # Use the low-level webapi functions directly to avoid recursion
 
@@ -1077,6 +999,9 @@ class Batch(WebContainer):
         """
         self._check_path_dir(path_dir=path_dir)
         self.download(path_dir=path_dir, replace_existing=replace_existing)
+
+        # Import BatchData here to avoid circular imports
+        from tidy3d.web.api.batch_data import BatchData
 
         if self.jobs is None:
             raise DataError("Can't load batch results, hasn't been uploaded.")
